moving logging config out of module based setup
